fix(tests): moved incorrect registered test to correct location
